Fix calculating decimals and numbers greater than 9

+++ calculate.h
@@ -1,1 +1,1 @@
-int calculate(stack input);
+double calculate(stack input);
